Refactored projection and it should be proper instead of relying on pizel size, seperated Sprite Renderer into quad and line to make modules specific to their role
