+++ test/inc/test_window.hpp
@@ -4,7 +4,7 @@
 #define TEST_WINDOW_HPP
 
 #include <cameras/ortho_camera.hpp>
-#include <ecs/systems/ecs_sprite_renderer.hpp>
+#include <ecs/systems/ecs_quad_renderer.hpp>
 #include <window/sisters_sdl_window.hpp>
 #include <input/sisters_sdl_gamepad.hpp>
 #include <input/managers/sisters_sdl_gamepad_manager.hpp>
@@ -14,7 +14,7 @@ using namespace SDL;
 class TestWindow : public Window {
     private:
         OrthoCamera camera;
-        ECS_SpriteRendererPtr renderer;
+        ECS_QuadRendererPtr renderer;
         Entity redEntity;
         Entity greenEntity;
     public:
